Changed project structure. Code is now in a package called production.
