+++ .gitignore
@@ -3,7 +3,7 @@
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
 
-# Package Files #
+# production.Package Files #
 *.jar
 *.war
 *.ear
