+++ src/production/EventConsumer.java
@@ -1,3 +1,5 @@
+package production;
+
 public interface EventConsumer {
     void handleTaskEvent(Task task, Event event);
 }
