+++ src/production/InventoryList.java
@@ -1,3 +1,5 @@
+package production;
+
 /**
  * Created by kyle on 11/10/16.
  */
@@ -139,7 +141,7 @@ public class InventoryList {
             new InventoryItem(876529,"First Aid Kit"),
             new InventoryItem(641348,"Fish Turner"),
             new InventoryItem(375111,"Fishing Reel"),
-            new InventoryItem(738632,"Floor Mat"),
+            new InventoryItem(738632,"production.Floor Mat"),
             new InventoryItem(253257,"Floorstanding Speaker"),
             new InventoryItem(155989,"Foam Roller"),
             new InventoryItem(977238,"Foldable Laundry Bin"),
@@ -183,7 +185,7 @@ public class InventoryList {
             new InventoryItem(125361,"Hanging Pot Rack"),
             new InventoryItem(545211,"Heat-Resistant Utensil Set"),
             new InventoryItem(690622,"Heated Mattress Pad"),
-            new InventoryItem(160200,"Hidden Book Shelf"),
+            new InventoryItem(160200,"Hidden Book production.Shelf"),
             new InventoryItem(143817,"Home Theater Projector"),
             new InventoryItem(147914,"Home Theater Speakers"),
             new InventoryItem(370123,"Huge Portable Hammock"),
@@ -292,7 +294,7 @@ public class InventoryList {
             new InventoryItem(714476,"Pressure Cooker"),
             new InventoryItem(585856,"Pressure Cooker  Canner"),
             new InventoryItem(719603,"Printer / Scanner / Copier / Fax"),
-            new InventoryItem(516855,"Programmable Lego Robot"),
+            new InventoryItem(516855,"Programmable Lego production.Robot"),
             new InventoryItem(228091,"Programmable Switch"),
             new InventoryItem(666364,"Propulsion Scooter"),
             new InventoryItem(793345,"Pulse Oximeter"),
