+++ src/production/InventoryManagement.java
@@ -1,3 +1,5 @@
+package production;
+
 import java.awt.*;
 import java.util.*;
