+++ src/production/InventoryManagementTest.java
@@ -1,3 +1,5 @@
+package production;
+
 import org.junit.Test;
 
 import java.awt.*;
