+++ src/production/Task.java
@@ -1,8 +1,10 @@
+package production;
+
 import java.awt.*;
 
 /**
  * A simple description of the action that will take place, together
- * with the EventConsumer that will handle it. This is essentially
+ * with the production.EventConsumer that will handle it. This is essentially
  * a struct
  */
 public class Task {
